Change routing structure
